Big changes, small changes.  Mostly clean ups.  One or two bugs.
BaseDisplay gave a lot of its functionality to Blackbox.
Basemenu now uses find() whenever it needs to find a member, other assorted
menu cleanups.
screen->toggleFocus() added.  The goal was to prevent a system wide reconfigure
for a simple change in focus model.
moved bexec to Util, all emx functionality is now in Util
basename is now in Util as well

general: unsigned and const cleanups, more work to be done here.
         most blackbox classes can not be copied, made this more clear by
           placing the copy constructor and op= in private sections
         all functions in headers now give argument names as well as type
         a few more uses of string over char*, more work here too

this adds one HUGE bug though -- changing Styles leaves the style menu open
and any attempt to close it kills blackbox with a sig11.