convert more parts of BlackboxWindow over to Rects... i also removed the y_border and y_handle stuff, and added a margin strut.  this makes it 10 times easier to calculate the frame rect from the client rect (and the other way), which makes the maximize/resize/upsize/downsize/positionWindows/etc. all much simpler.
I also changed frame.rect to hold the total size of the window (the full width
of the window including the border around the window). updated the smart
placement code to use this instead.  the framerect is also changed when
shading a window, so there is no need to check for this in the smart placement
code.

all in all, things are very good... window menu placement needs some work,
which i will do this weekend

OH, this change was prompted by my finding an off by one error while resizing
windows.  resize an xclock to 200x200, and it really gets the size of 201x201.
this has also been fixed now.