here's a nice diff that fixes quite a few of the warts in CVS:
1. it fixes a problem where acroread dialogs will disappear if you start a
find and switch workspaces away and back before the find is finished.

2. reintroduce server grabs around all reparenting and input mask changes.
this does add a slight bit of overhead (not noticible to me though), but
fixes most of the slitapp problems... i also stopped watching for
SubstructureNotify on the root windows... lots of extra events were being
generated, and in some cases we were getting 2 reparents every time we
managed a window... i don't know why i left it like this.  this is also
related to #1. above.

3. BlackboxWindow's constructor shouldn't take a default screen of 0.
BScreen calls this contructor, and the parameter is always passed...
remove the default value from the header, and remove the screen detection
code from the BlackboxWindow constructor.

4. introduce BWindowGroup (yeh yeh, i know, no new features, but the group
stuff that we had was severely broken).  this keeps all windows in a group
together, and allows you to get a non-transient window from the group
(VERY handy for the transient code, which also fixes those crazy group
transients).

5. root transients have their transient for set to ~0ul to keep
isTransient() returning true... this way, root transients still get focus
when they are shown (and they still look like transients).

6. fix BlackboxWindow::setInputFocus() for the GloballyActive case.
According to the ICCCM section 4.1.7, the window manager shouldn't set
input focus to the client window, BUT it should send the WM_TAKE_FOCUS
message. i also moved the autoraise timer starting/stopping into
setFocusFlag... it works best here (since we only want to raise windows
that get focus, the first time they get focus).

7. recent profiling shows that while starting up lots and lots of clients,
blackbox spends most of it's time constructing Rects (meaning, the
compiler generated code to allocate space on the heap/stack and call the
constructor, not the actual Rect constructors).  the smartplacement code
creates lots of temporary rects in loops.  we should create these once
outside the loop, and just set them up as we need them in the loop.  i
don't know if the compiler makes this kind of optimization, but i don't
profile with optimizations so it doesn't matter anyway :)

8. fix one special case where we get a MapRequest with a parent that isn't
root... very odd indeed, but something we need to make sure we don't fall
over on...