bug fixes:
shade a window and move it, the rubberband lines match the shaded size.  Now
change blackbox's style, hit reconfigure, whichever.  Now when you move the
shaded window the rubberband lines match the full size window.  This was
caused by the upsize() method.  We now check for shaded state and set the
height accordingly.

another shade problem.  shade a window, maximize it, then unmaximize it.  The
new window size is premax_h which was set to the height of the shaded window.
Similar to the fix for the problem above we now always use the client size for
this calculation.