Fixed the crash on workspace change bug.
The cause of the problem was blackbox attributes from a previous mapping
affecting the current mapping of a window.  This is the right approach during
startup because we want windows to return to their former state.  However this
is not the behaviour we want if a window is mapped and unmapped.

What was happening was this:

A window was mapped on workspace A and the unmapped.  It was then opened again
on workspace B.  During the managing of the window the blackbox attributes
from the last mapping were applied.  So the window believed it was on A while
it was actually on B.  If the user then switches to other workspaces the
window acts as if it were stuck because no workspace owns it.  On top of this
the window is marked as being the last focus window for workspace B.  Once
blackbox is on workspace A again the window is believed to be home.  If the
window is then unmapped the last focus pointer on workspace B is now invalid
but there is no way for blackbox to know this.  So the next time the current
workspace is changed to B a crash occurs as the invalid window pointer is
dereferenced.

The fix in this patch is to only restore window attributes during blackbox
startup.  As further protection against this bug I left an assert() in
BScreen::changeWorkspaceID().