revise netwm api:
the functions are now consistent in the sense that the first argument is the
window on which to store the property, followed by the data to set.  Also,
since we create the Netwm object with a display pointer, i have changed it
to store this pointer, and removed the Display * from all the functions.