lots of cleanups and fixes to make 8bpp work again.
for the moment, 8bpp default to ordered dither (this will be
configurable in the future).  this dither has been expanded to give a
bit better quality with a 16x16 dither matrix (instead of 8x8).

one big difference now is that we should be able to support ALL
possible color depths/visuals.  testing is needed to confirm this.

also, we no longer allocate the color cube at startup.  instead, we
just query the existing colors and use the closest matches we find.
this has actually proven to be much better than the old approach: my X
server comes up with the colormap already preallocated with the
standard web palette (6x6x6 color cube).  the image quality is great,
and we don't use up too many colors.  because of this, i have changed
the default colorsPerChannel to 6 now (instead of 4).  i will need to
test this on other 8bpp X servers (completeness).

all code using Pens now use the new Pen API (commit to follow).

oh, and i fixed the whiteline bug with crossdiagonal gradients :)