solve the last 8bpp issue with bsetroot.
add bsetroot::duplicateColor.  comment from the sources reads:

When using a colormap that is not read-only, we need to reallocate the
color we are using.  The application's color cache will be freed on
exit, and another client can allocate a new color at the same pixel
value, which will immediately change the color of the root window.

this is not what we want, so we need to make sure that the pixel we
are using is doubly allocated, so that it stays around with the
pixmap.  It will be released when we use
XKillClient(..., AllTemporary);