add Xft support to libbt...
since Xft cannot create fonts for use on all screens, we have to
change the API slightly for font-related things to take an 'screen'
parameter.

all in all, this is rather straight forward... we use the UTF-8
variants of the Xft API, but at the moment we have no real UTF-8
support.  we need to get this before we release 0.70.0