1. move the window and stacking lists from Workspace to BScreen.  this requires moving most of the functionality in Workspace to BScreen as well: - window placement - window stacking - focus fallback
2. introduce StackEntity and use it in the stacking list... this
allows us to have the Slit and Toolbar as proper stack participants.

3. finish the layered stacking support... this means that we can have
as many different layers as we want... currently we implement the
following:
  - fullscreen
  - always-on-top
  - normal
  - always-on-bottom
  - desktop

we now expose the always-on-top and always-on-bottom layers to the
end-user through the Windowmenu

4. implement smart full-screen support - when a focused full-screen
window is raised, it is placed above all other windows (including
windows marked always-on-top).  when it looses focus, it is moved to
the always-on-bottom layer.  raising the window will again bring it on
top of everything else.

5. cleanup BlackboxWindow and BScreen API to be more concise, easier
to understand and use.

6. many bugs and cleanups are also included in this commit... they
were a necessary part of completing the above changes and making sure
the new functionality is working